Refactor LieuHebergement constructor
This commit modifies the LieuHebergement class constructor to include a new parameter called chambres. This parameter is a list of Chambre objects and allows for the addition of multiple chambres to a LieuHebergement instance in a single constructor call. The previous implementation required the use of a separate method (ajouterChambre) to add chambres one by one. This change improves the usability and readability of the code.

Refactor LieuHebergementTest setUp method

This commit updates the setUp method in the LieuHebergementTest class. It now initializes the LieuHebergement instance with chambres directly in the constructor call, instead of using the ajouterChambre method. This change simplifies the test setup and makes it more concise.